more packet handlers
move packet + join packet handler + update players + ping packet

+++ src/networking/helpers.go
@@ -123,7 +123,7 @@ func appendString(list *[]byte, str string) {
 	*list = append(*list, data...)
 }
 
-func appendPosition(list *[]byte, x int16, y int16) {
+func appendPosition(list *[]byte, x uint16, y uint16) {
 	buf := make([]byte, 4)
 	binary.BigEndian.PutUint16(buf[0:2], uint16(x))
 	binary.BigEndian.PutUint16(buf[2:4], uint16(y))
@@ -167,12 +167,12 @@ func readString(data []byte, offset *int) (string, error) {
 	return str, nil
 }
 
-func readPosition(data []byte, offset *int) (int16, int16, error) {
+func readPosition(data []byte, offset *int) (uint16, uint16, error) {
 	if *offset + 4 > len(data) {
 		return 0, 0, fmt.Errorf("position: not enough data")
 	}
-	x := int16(binary.BigEndian.Uint16(data[*offset : *offset+2]))
-	y := int16(binary.BigEndian.Uint16(data[*offset+2 : *offset+4]))
+	x := binary.BigEndian.Uint16(data[*offset : *offset+2])
+	y := binary.BigEndian.Uint16(data[*offset+2 : *offset+4])
 	*offset += 4
 	return x, y, nil
 }
